Fix access warning by moving generated class and making it package private
