+++ org.eclipse.january/META-INF/MANIFEST.MF
@@ -6,7 +6,6 @@ Bundle-Version: 2.0.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.eclipse.january,
  org.eclipse.january.dataset,
- org.eclipse.january.dataset.internal;x-internal:=true,
  org.eclipse.january.io,
  org.eclipse.january.metadata,
  org.eclipse.january.metadata.internal;x-internal:=true
