+++ org.eclipse.january/src/org/eclipse/january/dataset/GeneratedMaths.java
@@ -10,34 +10,14 @@
  *    Peter Chang - initial API and implementation and/or initial documentation
  *******************************************************************************/
 
-package org.eclipse.january.dataset.internal;
+package org.eclipse.january.dataset;
 
 import org.apache.commons.math3.complex.Complex;
-import org.eclipse.january.dataset.BroadcastIterator;
-import org.eclipse.january.dataset.ByteDataset;
-import org.eclipse.january.dataset.ComplexDoubleDataset;
-import org.eclipse.january.dataset.ComplexFloatDataset;
-import org.eclipse.january.dataset.CompoundByteDataset;
-import org.eclipse.january.dataset.CompoundDoubleDataset;
-import org.eclipse.january.dataset.CompoundFloatDataset;
-import org.eclipse.january.dataset.CompoundIntegerDataset;
-import org.eclipse.january.dataset.CompoundLongDataset;
-import org.eclipse.january.dataset.CompoundShortDataset;
-import org.eclipse.january.dataset.DTypeUtils;
-import org.eclipse.january.dataset.Dataset;
-import org.eclipse.january.dataset.DatasetFactory;
-import org.eclipse.january.dataset.DoubleDataset;
-import org.eclipse.january.dataset.FloatDataset;
-import org.eclipse.january.dataset.IntegerDataset;
-import org.eclipse.january.dataset.LongDataset;
-import org.eclipse.january.dataset.Operations;
-import org.eclipse.january.dataset.ShortDataset;
-import org.eclipse.january.dataset.SingleInputBroadcastIterator;
 
 /**
  * Generated by template/generatefunctions.py
  */
-public class GeneratedMaths {
+class GeneratedMaths {
 	protected static void addFunctionName(final Dataset a, final Dataset b, final Dataset dataset, final String fname) {
 		dataset.setName(Operations.createFunctionName(fname, a.getName(), b.getName()));
 	}
