+++ org.eclipse.january/src/org/eclipse/january/dataset/internal/template/generatefunctions.py
@@ -17,7 +17,7 @@ Takes a functions definition file and the shell of a Java class,
 generates the methods and prints to standard output the completed class
 
 Runs as follows
-$ python generatefunctions.py functions.txt ../GeneratedMaths.java > GeneratedMaths.java
+$ python generatefunctions.py functions.txt ../../GeneratedMaths.java > GeneratedMaths.java
 
 The format is
